patient_service/requirements.txt
Django>=5.0
djangorestframework
gunicorn
mysqlclient>=2.2
django-cors-headers==4.4.0
requests==2.32.3
djangorestframework-simplejwt==5.3.1